Pratical Exam Paper for DM - T2206M
A-Z bank needed to find a more flexible infrastructure for its core banking to make it easier
to integrate and deploy new solutions to help it keep pace with its rapid growth. A-Z Bank
plans to use SQL Server 2005 Database Mirroring to enhance high availability of its data.
And you are required to create a database name AZBank using following sql statements.
The diagram of AZBank as shown following.